requirements.txt
neo4j>=5.20.0
supabase>=2.4.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.5.0
httpx>=0.27.0
litellm>=1.48.0
langgraph>=0.1.0
langchain-core>=0.2.0
